[streaming] [scala] Restructured streaming scala project and examples
Amended for the release-0.8 branch
